Make the panic handling functionality public
I think I will need this for uniffi#460.  It seems reasonable that if
ffi-support is going to set the global panic handling hook, then it
should give other crates the opportunity to initialize it if they need
to.

Updated the name to something that hopefully describes the intent a bit
more.